Redux store set up wrapping gatsby root.
