+++ gatsby-ssr.js
@@ -1,7 +1,2 @@
-/**
- * Implement Gatsby's SSR (Server Side Rendering) APIs in this file.
- *
- * See: https://www.gatsbyjs.org/docs/ssr-apis/
- */
-
-// You can delete this file if you're not using it
+import wrapWithProvider from "./wrap-with-provider"
+export const wrapRootElement = wrapWithProvider
